tests/blackbox/pyproject.toml
[tool.poetry]
name = "Kyma Companion Blackbox Tests"
version = "0.1.0"
description = "Evaluation and Validation Framework for Kyma Companion"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "~3.12"

python-dotenv = "^1.0.1"
pydantic = "^2.8.2"
termcolor = "^2.4.0"
generative-ai-hub-sdk = { version = "^1.11.1", extras = ["langchain"] }
prettytable = "^3.10.2"
wcwidth = "^0.2.13"
